% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_step_1.5_plot_baseline_sales.R
\name{optional_step_1.5_plot_baseline_sales}
\alias{optional_step_1.5_plot_baseline_sales}
\title{(Optional) Step 1.5: Plot Graph of Baseline Sales}
\usage{
optional_step_1.5_plot_baseline_sales()
}
\value{
A ggplot object
}
\description{
Plots baseline sales simulated from step 1.
}
\examples{
\dontrun{
optional_step_1.5_plot_baseline_sales()
}
}
