% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_step_2.5_plot_ad_spend.R
\name{optional_step_2.5_plot_ad_spend}
\alias{optional_step_2.5_plot_ad_spend}
\title{(Optional) Step 2.5: Plot Graph of Ad Spend}
\usage{
optional_step_2.5_plot_ad_spend()
}
\value{
A ggplot object
}
\description{
Plots ad spend simulated from step 2.
}
\examples{
\dontrun{
optional_step_2.5_plot_ad_spend()
}
}
