% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_2_ads_spend.R
\name{step_2_ads_spend}
\alias{step_2_ads_spend}
\title{Step 2: Generate Ad Spend}
\usage{
step_2_ads_spend(
  campaign_spend_mean,
  campaign_spend_std,
  max_min_proportion_on_each_channel
)
}
\arguments{
\item{campaign_spend_mean}{A numeric, the average amount of money spent on a campaign, must be in same currency as baseline sales generated in step 1}

\item{campaign_spend_std}{A numeric, the standard deviation of money spent on a campaign, must be in same currency as baseline sales generated in step 1}

\item{max_min_proportion_on_each_channel}{A vector of numerics specifying the minimum and maximum percentages of total spend allocated to each channel, should be in the same order as channels specified (channels that use impressions first followed by channels that use clicks), length should be 2 times (number of total channels - 1)}
}
\value{
A data frame with amounts of money spent on each campaign and media channel
}
\description{
Simulates how much to spend on each ad campaign and channel
}
\examples{
\dontrun{
step_2_ads_spend(campaign_spend_mean = 329000,
campaign_spend_std = 100000,
max_min_proportion_on_each_channel <- c(0.45, 0.55,
0.15, 0.25,
0.1, 0.2)
)
}
}
