% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_3_generate_media.R
\name{step_3_generate_media}
\alias{step_3_generate_media}
\title{Step 3 : Generate Media Variables}
\usage{
step_3_generate_media(
  true_cpm,
  true_cpc,
  mean_noisy_cpm_cpc,
  std_noisy_cpm_cpc
)
}
\arguments{
\item{true_cpm}{A vector of numbers specifying the true Cost per Impression (CPM) of each channel (noise will be added to this to simulate number of impressions), length MUST equal to the number of TOTAL channels, if channels do not use impressions to measure activity, then put NA, must be in same order as channels specified (put channels that use impressions first, followed by channels that use clicks)}

\item{true_cpc}{A vector of numbers specifying the true Cost per Click (CPC) of each channel (noise will be added to this to simulate number of clicks), length MUST equal to the number of TOTAL channels, if channels do not use clicks to measure activity, then put NA, must be in same order as channels specified (put channels that use impressions first, followed by channels that use clicks)}

\item{mean_noisy_cpm_cpc}{A vector of numbers with mean of normal distribution that generates noise to CPM or CPC, vector with length equal to number of channels, must be in same order as channels specified (put channels that use impressions first, followed by channels that use clicks)}

\item{std_noisy_cpm_cpc}{A vector of numbers with standard deviation of normal distribution that generates noise to CPM or CPC, vector with length equal to number of channels, must be in same order as channels specified (put channels that use impressions first, followed by channels that use clicks)}
}
\value{
A data frame with activity for media channels
}
\description{
This step generates media activity using spend from Step 2.
}
\examples{
\dontrun{

step_3_generate_media(
true_cpm = c(2, 20, 10, NA),
true_cpc = c(NA, NA, NA, 1),
mean_noisy_cpm_cpc = c(0, 0.05, 0.1, 0.02),
std_noisy_cpm_cpc = c(0.1, 0.15, 0.2, 0.03)
)
}
}
