% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_5a_pivot_to_mmm_format.R
\name{step_5a_pivot_to_mmm_format}
\alias{step_5a_pivot_to_mmm_format}
\title{Step 5a: Pivot from Our campaign + channel format to a daily format for MMM format}
\arguments{
\item{my_variables}{A list that was created after running step 0. It stores the inputs you've specified.}

\item{df_ads_step4}{A data frame that was created after running step 4.}
}
\value{
A data frame with media variables, media spend
}
\description{
We pivot our table from Step 4 (which was on a campaign + channel format) to a daily format.(We will create a weekly format later.)
}
\examples{
\dontrun{
step_5a_pivot_to_mmm_format(
my_variables = my_variables,
df_ads_step4 = df_ads_step4,
}
}
