% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_6_calculating_conversions.R
\name{step_6_calculating_conversions}
\alias{step_6_calculating_conversions}
\title{Step 6 : Calculating Conversions}
\usage{
step_6_calculating_conversions(
  my_variables = my_variables,
  df_ads_step5c = df_ads_step5c
)
}
\arguments{
\item{my_variables}{A list that was created after running step 0. It stores the inputs you've specified.}

\item{df_ads_step5c}{A data frame that was created after running step 5c.}
}
\value{
A data frame with number of conversions on each channel
}
\description{
We calculate the number of conversions that occur each week and on each channel using the outputs from previous steps.
}
\examples{
\dontrun{
step_6_calculating_conversions(
  my_variables = my_variables,
  df_ads_step5c = df_ads_step5c
  )
}

}
