% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_7_expanded_df.R
\name{step_7_expanded_df}
\alias{step_7_expanded_df}
\title{Step 7 : Expanded Data Frame}
\usage{
step_7_expanded_df(
  my_variables = my_variables,
  df_ads_step6 = df_ads_step6,
  df_baseline = df_baseline
)
}
\arguments{
\item{my_variables}{A list that was created after running step 0. It stores the inputs you've specified.}

\item{df_ads_step6}{A data frame that was created after running step 6.}

\item{df_baseline}{A data frame that was created after running step 1.}
}
\value{
A data frame with all the variables we've calculated so far.
}
\description{
We do final calculations for all the variables that we need for the final data frame.
}
\examples{
 \dontrun{
step_7_expanded_df(
  my_variables = my_variables,
  df_ads_step6 = df_ads_step6,
  df_baseline = df_baseline
  )
}
}
