% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_8_calculate_roi.R
\name{step_8_calculate_roi}
\alias{step_8_calculate_roi}
\title{Step 8 : Calculating ROI}
\usage{
step_8_calculate_roi()
}
\value{
A list of ROIs for each channel
}
\description{
We calculate ROIs of each channel. These are the 'true ROIs' that we generated the data set with. We can use these to compare against the ROIs predicted from our MMM and see how close our MMM comes to these true ROIs.
}
\examples{
\dontrun{
step_8_calculate_roi()
}
}
