% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optional_step_9.5_plot_final_df.R
\name{optional_step_9.5_plot_final_df}
\alias{optional_step_9.5_plot_final_df}
\title{(Optional) Step 9.5: Plot Graph of Final Data Set}
\usage{
optional_step_9.5_plot_final_df(df_final = df_final)
}
\arguments{
\item{df_final}{Your final data frame that came from Step 9.}
}
\value{
A ggplot object
}
\description{
Plots final data set with date on x-axis and revenue on y-axis.
}
\examples{
\dontrun{
optional_step_9.5_plot_final_df(
df_final = df_final
)
}
}
