% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_4_generate_cvr.R
\name{step_4_generate_cvr}
\alias{step_4_generate_cvr}
\title{Step 4 : Generate Noisy CVRs}
\usage{
step_4_generate_cvr(
  my_variables = my_variables,
  df_ads_step3 = df_ads_step3,
  mean_noisy_cvr,
  std_noisy_cvr
)
}
\arguments{
\item{my_variables}{A list that was created after running step 0. It stores the inputs you've specified.}

\item{df_ads_step3}{A data frame that was created after running step 3.}

\item{mean_noisy_cvr}{A vector of numbers specifying the mean of the normal distribution used to add noise to conversion rates (CVR) of each channel, vector must be in the same order as channels specified (first channels that use impressions as a metric of activity, then channels that use clicks), must have length equal to number of total channels}

\item{std_noisy_cvr}{A vector of numbers specifying the standard deviation of the normal distribution used to add noise to conversion rates (CVR) of each channel, vector must be in the same order as channels specified (first channels that use impressions as a metric of activity, then channels that use clicks), must have length equal to number of total channels}
}
\value{
A data frame with noisy conversion rates for each channel and campaign
}
\description{
Generate noise around CVRs that user inputs for each channel
}
\examples{
\dontrun{
step_4_generate_cvr(
my_variables = my_variables,
df_ads_step3 = df_ads_step3,
mean_noisy_cvr = c(0, 0.0001, 0.0002, 0.0005),
std_noisy_cvr = c(0.001, 0.002, 0.0001, 0.006))
}
}
