% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_5a_pivot_to_mmm_format.R
\name{step_5a_pivot_to_mmm_format}
\alias{step_5a_pivot_to_mmm_format}
\title{Step 5a: Pivot from Our campaign + channel format to a weekly format}
\value{
A data frame with media variables, media spend
}
\description{
We pivot our table from Step 4 (which was on a campaign + channel format) to a weekly format. This function requires no parameters.
}
\examples{
\dontrun{
step_5a_pivot_to_mmm_format()
}
}
