% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_5b_decay.R
\name{step_5b_decay}
\alias{step_5b_decay}
\title{Step 5b: Apply Adstock Decay}
\usage{
step_5b_decay(
  my_variables = my_variables,
  df_ads_step5a_before_mmm = df_ads_step5a_before_mmm,
  true_lambda_decay
)
}
\arguments{
\item{my_variables}{A list that was created after running step 0. It stores the inputs you've specified.}

\item{df_ads_step5a_before_mmm}{A data frame that was created after running step 5a.}

\item{true_lambda_decay}{A vector of numbers between 0 and 1 specifying the lambda parameters for a geometric distribution for adstocking media variables. Vector must be in same order that channels are specified (channels using impressions as a metric followed by channels using click as metric of activity), must have same length as total number of channels}
}
\value{
A data frame with adstocked media variables
}
\description{
Apply adstock decay (using geometric distribution) to media variables
}
\examples{
\dontrun{
step_5b_decay(
 my_variables = my_variables,
 df_ads_step5a_before_mmm = df_ads_step5a_before_mmm,
true_lambda_decay = c(0.1, 0.2, 0.15, 0.08)
)
}
}
