% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_6_calculating_conversions.R
\name{step_6_calculating_conversions}
\alias{step_6_calculating_conversions}
\title{Step 6 : Calculating Conversions}
\usage{
step_6_calculating_conversions()
}
\value{
A data frame with number of conversions on each channel
}
\description{
We calculate the number of conversions that occur each week and on each channel using the outputs from previous steps.
}
\examples{
\dontrun{
step_6_calculating_conversions()
}

}
