% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_7_expanded_df.R
\name{step_7_expanded_df}
\alias{step_7_expanded_df}
\title{Step 7 : Expanded Data Frame}
\usage{
step_7_expanded_df()
}
\value{
A data frame with all the variables we've calculated so far.
}
\description{
We do final calculations for all the variables that we need for the final data frame.
}
\examples{
 \dontrun{
step_7_expanded_df()
}
}
