% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_9_final_df.R
\name{step_9_final_df}
\alias{step_9_final_df}
\title{Step 9 : Get Final Data Frame}
\usage{
step_9_final_df()
}
\value{
A data frame to use to put into an MMM
}
\description{
Keep only necessary variables for final data frame and format
}
\examples{
\dontrun{
step_9_final_df()
}
}
